requirements.txt
numpy==1.24.1
pandas==2.0.3
torch==1.11.0
#transformers @ git+https://github.com/huggingface/transformers@7790943c91411f4234d11dfbf4c2f21ce7caf088
requests==2.28.1
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
googletrans==3.1.0a0
flask==3.0.0
fastapi==0.109.0
tqdm==4.66.1
urllib3==1.26.13
gdown==4.7.1
langchain==0.0.320
langsmith==0.0.80 
langchain-core==0.1.10 
openai==0.28.0
openpyxl==3.1.2
gdown==4.7.1
transformers
tqdm==4.66.1
tqdm-multiprocess==0.0.11
regex==2023.8.8
json5==0.9.14
jsonlines==4.0.0
huggingface-hub==0.16.4
python-dotenv==1.0.1
isodate==0.6.1
#llama-index
